+++ packages/kit/src/views/Market/components/MarketDetailPools.tsx
@@ -1,4 +1,3 @@
-import type { PropsWithChildren } from 'react';
 import { useCallback, useMemo, useState } from 'react';
 
 import { useIntl } from 'react-intl';
@@ -9,7 +8,7 @@ import {
   Icon,
   NumberSizeableText,
   SizableText,
-  Stack,
+  Skeleton,
   Table,
   View,
   XStack,
@@ -20,6 +19,7 @@ import {
 import { useSettingsPersistAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
 import type {
+  IMarketDetailPlatform,
   IMarketDetailPool,
   IMarketDetailTicker,
   IMarketResponsePool,
@@ -91,17 +91,88 @@ function HeaderColumn({
   );
 }
 
+function SkeletonRow() {
+  return (
+    <XStack>
+      <XStack flex={1}>
+        <Skeleton w="$24" h="$3" />
+      </XStack>
+      <XStack flex={1} jc="flex-end">
+        <Skeleton w="$16" h="$3" />
+      </XStack>
+      <XStack flex={1} jc="flex-end">
+        <Skeleton w="$16" h="$3" />
+      </XStack>
+      <XStack flex={1} jc="flex-end">
+        <Skeleton w="$16" h="$3" />
+      </XStack>
+      <XStack flex={1} jc="flex-end">
+        <Skeleton w="$16" h="$3" />
+      </XStack>
+    </XStack>
+  );
+}
+
+function MdSkeletonRow() {
+  return (
+    <XStack>
+      <XStack flex={1}>
+        <Skeleton w="$24" h="$3" />
+      </XStack>
+      <XStack flex={1} jc="flex-end">
+        <Skeleton w="$16" h="$3" />
+      </XStack>
+      <XStack flex={1} jc="flex-end">
+        <Skeleton w="$16" h="$3" />
+      </XStack>
+    </XStack>
+  );
+}
+
+function MarketDetailPoolsSkeletonRow() {
+  const { md } = useMedia();
+  return md ? (
+    <YStack gap="$10" px="$5" pt="$11">
+      <MdSkeletonRow />
+      <MdSkeletonRow />
+      <MdSkeletonRow />
+      <MdSkeletonRow />
+    </YStack>
+  ) : (
+    <YStack gap="$6" px="$5" pt="$11">
+      <SkeletonRow />
+      <SkeletonRow />
+      <SkeletonRow />
+      <SkeletonRow />
+    </YStack>
+  );
+}
+
 export function MarketDetailPools({
-  pools,
+  detailPlatforms,
   tickers,
 }: ITabPageProps & {
-  pools: IMarketResponsePool[];
   tickers?: IMarketDetailTicker[];
+  detailPlatforms: IMarketDetailPlatform;
 }) {
   const [settings] = useSettingsPersistAtom();
   const currency = settings.currencyInfo.symbol;
   const intl = useIntl();
   const { gtXl } = useMedia();
+
+  const { result: pools } = usePromiseResult(
+    async () => {
+      const response = await backgroundApiProxy.serviceMarket.fetchPools(
+        detailPlatforms,
+      );
+      return response;
+    },
+    [detailPlatforms],
+    {
+      initResult: [],
+    },
+  );
+
   const oneKeyNetworkIds = useMemo(() => {
     const result = pools
       .map((i) => i.onekeyNetworkId)
@@ -497,6 +568,7 @@ export function MarketDetailPools({
           onChange={handleChange}
         />
       }
+      TableEmptyComponent={<MarketDetailPoolsSkeletonRow />}
       renderScrollComponent={renderNestedScrollView}
       onRow={onRow}
       onHeaderRow={onHeaderRow as any}
