+++ packages/kit/src/views/Market/components/TokenDetailTabs.tsx
@@ -34,44 +34,6 @@ import { TokenPriceChart } from './TokenPriceChart';
 import type { IDeferredPromise } from '../../../hooks/useDeferredPromise';
 import type { LayoutChangeEvent } from 'react-native';
 
-function SkeletonRow() {
-  return (
-    <XStack>
-      <XStack flex={1}>
-        <Skeleton w="$24" h="$3" />
-      </XStack>
-      <XStack flex={1} jc="flex-end">
-        <Skeleton w="$16" h="$3" />
-      </XStack>
-      <XStack flex={1} jc="flex-end">
-        <Skeleton w="$16" h="$3" />
-      </XStack>
-      <XStack flex={1} jc="flex-end">
-        <Skeleton w="$16" h="$3" />
-      </XStack>
-      <XStack flex={1} jc="flex-end">
-        <Skeleton w="$16" h="$3" />
-      </XStack>
-    </XStack>
-  );
-}
-
-function MdSkeletonRow() {
-  return (
-    <XStack>
-      <XStack flex={1}>
-        <Skeleton w="$24" h="$3" />
-      </XStack>
-      <XStack flex={1} jc="flex-end">
-        <Skeleton w="$16" h="$3" />
-      </XStack>
-      <XStack flex={1} jc="flex-end">
-        <Skeleton w="$16" h="$3" />
-      </XStack>
-    </XStack>
-  );
-}
-
 function BasicTokenDetailTabs({
   token,
   listHeaderComponent,
@@ -90,109 +52,65 @@ function BasicTokenDetailTabs({
   const intl = useIntl();
   const { md } = useMedia();
 
-  const [pools, setPools] = useState<
-    | {
-        data: IMarketDetailPool[];
-        contract_address: string;
-        onekeyNetworkId?: string | undefined;
-        coingeckoNetworkId?: string | undefined;
-      }[]
-    | undefined
-  >(undefined);
-
-  const init = useCallback(async () => {
-    if (token?.detailPlatforms) {
-      const response = await backgroundApiProxy.serviceMarket.fetchPools(
-        token.detailPlatforms,
-      );
-
-      setTimeout(() => {
-        defer.resolve(null);
-      }, 100);
-      setPools(response);
-    }
-  }, [defer, token?.detailPlatforms]);
-  useEffect(() => {
-    void init();
-  }, [init]);
-
-  const renderPoolSkeleton = useMemo(
-    () =>
-      md ? (
-        <YStack gap="$10" px="$5" pt="$11">
-          <MdSkeletonRow />
-          <MdSkeletonRow />
-          <MdSkeletonRow />
-          <MdSkeletonRow />
-        </YStack>
-      ) : (
-        <YStack gap="$6" px="$5" pt="$11">
-          <SkeletonRow />
-          <SkeletonRow />
-          <SkeletonRow />
-          <SkeletonRow />
-        </YStack>
-      ),
-    [md],
-  );
+  setTimeout(() => {
+    defer.resolve(null);
+  }, 100);
 
   const tabConfig = useMemo(
     () =>
-      pools
-        ? [
-            md && token
-              ? {
-                  title: intl.formatMessage({
-                    id: ETranslations.global_overview,
-                  }),
-                  // eslint-disable-next-line react/no-unstable-nested-components
-                  page: (props: ITabPageProps) => (
-                    <TokenPriceChart
-                      {...props}
-                      tickers={token?.tickers}
-                      coinGeckoId={coinGeckoId}
-                      defer={defer}
-                      symbol={token?.symbol}
-                    />
-                  ),
-                }
-              : undefined,
-            md && token
-              ? {
-                  title: intl.formatMessage({
-                    id: ETranslations.global_overview,
-                  }),
-                  // eslint-disable-next-line react/no-unstable-nested-components
-                  page: (props: ITabPageProps) => (
-                    <MarketDetailOverview {...props} token={token} />
-                  ),
-                }
-              : undefined,
-            (pools.length || token?.tickers?.length) && token
-              ? {
-                  title: intl.formatMessage({ id: ETranslations.global_pools }),
-                  // eslint-disable-next-line react/no-unstable-nested-components
-                  page: (props: ITabPageProps) => (
-                    <MarketDetailPools
-                      {...props}
-                      pools={pools}
-                      tickers={token.tickers}
-                    />
-                  ),
-                }
-              : undefined,
-            token && {
+      [
+        md && token
+          ? {
               title: intl.formatMessage({
-                id: ETranslations.global_links,
+                id: ETranslations.global_overview,
               }),
               // eslint-disable-next-line react/no-unstable-nested-components
               page: (props: ITabPageProps) => (
-                <MarketDetailLinks {...props} token={token} />
+                <TokenPriceChart
+                  {...props}
+                  tickers={token?.tickers}
+                  coinGeckoId={coinGeckoId}
+                  defer={defer}
+                  symbol={token?.symbol}
+                />
               ),
-            },
-          ].filter(Boolean)
-        : [],
-    [coinGeckoId, defer, intl, md, pools, token],
+            }
+          : undefined,
+        md && token
+          ? {
+              title: intl.formatMessage({
+                id: ETranslations.global_overview,
+              }),
+              // eslint-disable-next-line react/no-unstable-nested-components
+              page: (props: ITabPageProps) => (
+                <MarketDetailOverview {...props} token={token} />
+              ),
+            }
+          : undefined,
+        token?.tickers?.length && token
+          ? {
+              title: intl.formatMessage({ id: ETranslations.global_pools }),
+              // eslint-disable-next-line react/no-unstable-nested-components
+              page: (props: ITabPageProps) => (
+                <MarketDetailPools
+                  {...props}
+                  tickers={token.tickers}
+                  detailPlatforms={token.detailPlatforms}
+                />
+              ),
+            }
+          : undefined,
+        token && {
+          title: intl.formatMessage({
+            id: ETranslations.global_links,
+          }),
+          // eslint-disable-next-line react/no-unstable-nested-components
+          page: (props: ITabPageProps) => (
+            <MarketDetailLinks {...props} token={token} />
+          ),
+        },
+      ].filter(Boolean),
+    [coinGeckoId, defer, intl, md, token],
   );
 
   const tabRef = useRef<ITabInstance | null>(null);
